returns temp in celcius now

+++ NILE_Hardware/adc_i2c.h
@@ -1,3 +1,5 @@
+//Driver for AD7995 4-Channel 10 Bit ADC
+//Nicodemus Phaklides
 #ifndef adc_i2c
 #define adc_i2c
 
@@ -8,7 +10,7 @@ class ADCI2C
 public:
   ADCI2C(int address);
   void init();
-  int read(int channel);
+  double read(int channel);
 private:
   int adr;
 };
